requirements.txt
# Flask核心依赖
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# 数据库驱动
psycopg2-binary==2.9.9
PyMySQL==1.1.0
redis==5.0.1

# 数据处理和验证
marshmallow==3.20.1
python-dotenv==1.0.0

# 生产环境服务器
gunicorn==21.2.0

# 开发和测试工具
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# 监控和日志
flask-limiter==3.5.0
